--- cms_app/src/index.css
@@ -1,33 +1,0 @@
-html {
-  font-size: 16px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-}
-
-body {
-  /* border: black 2px solid; */
-  margin-top: 20px;
-  aspect-ratio: 5 /2;
-  width: 80%;
-  background-color: #74c5ab;
-
-  
-  
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-::-webkit-scrollbar {
-  display: none; 
-  /* hide scroll bar to avoid display ratio confusion */
-}
